Updates tests with cancellation token.
