+++ tests/NScatterGather.Tests/_TestTypes/SomeCancellableType.cs
@@ -7,7 +7,8 @@ namespace NScatterGather
     {
         public async Task<string> EchoAsString(int n, CancellationToken cancellationToken)
         {
-            await Task.Delay(1_000, cancellationToken);
+            try { await Task.Delay(500, cancellationToken); }
+            catch (TaskCanceledException) { }
             return n.ToString();
         }
     }
