+++ tests/NScatterGather.Tests/_TestTypes/SomeLongProcessingType.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+
+namespace NScatterGather
+{
+    public class SomeLongProcessingType
+    {
+        public async Task<string> Process(int n)
+        {
+            await Task.Delay(2_000);
+            return n.ToString();
+        }
+    }
+}
