+++ tests/NScatterGather.Tests/_Utils/CancellationHelpers.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace NScatterGather
+{
+    internal static class CancellationHelpers
+    {
+        public static async Task CatchCancellation(Task cancellableTask)
+        {
+            try { await cancellableTask; }
+            catch (OperationCanceledException) { }
+        }
+    }
+}
